feat(script/build): parallelize the compilation and optimization of models
